Find/Replace overlay: separate command infrastructure into dedicated class
Introduce FindReplaceOverlayCommandSupport to own the overlay's
command-related infrastructure. As a first step, move the workaround
that disables the target text editor's global action handlers
(setTextEditorActionsActivated and DeactivateGlobalActionHandlers)
out of FindReplaceOverlay into this new class. FindReplaceOverlay now
only calls commandSupport.onFocusGained/Lost().

Contributes to
https://github.com/eclipse-platform/eclipse.platform.ui/issues/1912
